+++ pkg/engine/charging/usage.go
@@ -2,10 +2,10 @@ package engine
 
 //Usage specifies the usage for a single service
 type Usage struct {
-	Volume        int64
+	Volume        float32
 	Unit          Unit
-	Charge        int64
-	Tax           int64
+	Charge        float32
+	Tax           float32
 	HomeTadigs    []string
 	VisitorTadigs []string
 }
@@ -24,7 +24,7 @@ type AggregatedUsage map[ServiceTadig]Usage
 // TODO take account of units and currency? (Not in MVP1)
 //Aggregate gets all the usage for every service / tadig combination and then aggregates them
 func (a AggregatedUsage) Aggregate(service string, htadigs []string, vtadigs []string) Usage {
-	var volume, charge, tax int64
+	var volume, charge, tax float32
 	for _, h := range htadigs {
 		for _, v := range vtadigs {
 			key := ServiceTadig{service, h, v}
